+++ modules/backend/database/migrations/2014_10_01_000007_Db_Backend_Add_Description_Field.php
@@ -1,0 +1,23 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class DbBackendAddDescriptionField extends Migration
+{
+    public function up()
+    {
+        Schema::table('backend_user_groups', function (Blueprint $table) {
+            $table->text('description')->nullable();
+            $table->boolean('is_new_user_default')->default(false);
+        });
+    }
+
+    public function down()
+    {
+        Schema::table('backend_user_groups', function (Blueprint $table) {
+            $table->dropColumn('description');
+            $table->dropColumn('is_new_user_default');
+        });
+    }
+}