+++ modules/system/database/migrations/2013_10_01_000007_Db_System_Add_Disabled_Flag.php
@@ -5,7 +5,6 @@ use Illuminate\Database\Migrations\Migration;
 
 class DbSystemAddDisabledFlag extends Migration
 {
-
     public function up()
     {
         Schema::table('system_plugin_versions', function (Blueprint $table) {